fix tests for templates and folder purging

+++ app/dao/templates_dao.py
@@ -14,7 +14,6 @@ from app.models import (
     Template,
     TemplateHistory,
     TemplateRedacted,
-    template_folder_map,
 )
 
 
@@ -170,11 +169,10 @@ def dao_purge_templates_for_service(service_id):
 
     ids = [f"'{str(x.id)}'" for x in templates]
     ids_string = ", ".join(ids)
-    query = f"DELETE FROM template_folder_map WHERE template_id IN ({ids_string})"
-    db.session.execute(query)
-    db.session.flush()
-
-    db.session.query(template_folder_map).all()
+    if len(ids_string) > 0:
+        query = f"DELETE FROM template_folder_map WHERE template_id IN ({ids_string})"
+        db.session.execute(query)
+        db.session.flush()
 
     messages_from_templates = BroadcastMessage.query.filter(
         BroadcastMessage.service_id == service_id, BroadcastMessage.template_id.isnot(None)
